Split up logic to match phone app
